move previous code to separate directory
