+++ packages/app-mobile/_proto/components/Image/index.tsx
