+++ packages/app-mobile/_proto/components/Path/index.tsx
