+++ packages/app-mobile/_proto/utils/selectElements.ts
